requirements.txt
Django>=4.2,<5.0
psycopg2-binary
djangorestframework
drf-spectacular
gunicorn
sqlalchemy
google-adk
python-dotenv
adrf
asyncpg